+++ src/Category_Colors/Admin.php
@@ -2,16 +2,14 @@
 
 namespace Fragen\Category_Colors;
 
-use Tribe__Events__Main,
-	Tribe__Settings_Tab;
-
+use Tribe__Events__Main;
+use Tribe__Settings_Tab;
 
 class Admin {
 
 	const TAB_NAME      = 'category-colors';
 	const UPDATE_ACTION = 'category-colors-update-options';
 	protected $teccc    = null;
-
 
 	public function __construct( Main $teccc ) {
 		$this->teccc = $teccc;
@@ -24,11 +22,9 @@ class Admin {
 		load_plugin_textdomain( 'the-events-calendar-category-colors', false, TECCC_LANG );
 	}
 
-
 	public function init() {
 		register_setting( 'teccc_category_colors', 'teccc_options', array( $this, 'validate_options' ) );
 	}
-
 
 	public function plugin_fail_msg() {
 		if ( current_user_can( 'activate_plugins' ) && is_admin() ) {
@@ -41,7 +37,6 @@ class Admin {
 			}
 		}
 	}
-
 
 	/**
 	 * @param array $input
@@ -86,13 +81,11 @@ class Admin {
 		return $input;
 	}
 
-
 	public function load_settings_tab() {
 		if ( class_exists( 'Tribe__Events__Main' ) ) {
 			add_action( 'tribe_settings_do_tabs', array( $this, 'add_category_colors_tab' ) );
 		}
 	}
-
 
 	public function add_category_colors_tab() {
 		$categoryColorsTab = $this->teccc->load_config( 'admintab' );
@@ -101,16 +94,13 @@ class Admin {
 		new Tribe__Settings_Tab( self::TAB_NAME, esc_html__( 'Category Colors', 'the-events-calendar-category-colors' ), $categoryColorsTab );
 	}
 
-
 	public function form_header() {
 		echo '<form method="post" action="options.php">';
 	}
 
-
 	public function settings_fields() {
 		settings_fields( 'teccc_category_colors' );
 	}
-
 
 	public function is_saved() {
 		if ( isset( $_GET['settings-updated'] ) && $_GET['settings-updated'] ) {
@@ -119,7 +109,6 @@ class Admin {
 			echo apply_filters( 'tribe_settings_success_message', $output, 'category-colors' );
 		}
 	}
-
 
 	public static function options_elements() {
 		$teccc = Main::instance();
@@ -133,7 +122,6 @@ class Admin {
 
 		return $content;
 	}
-
 
 	/**
 	 * Retrieves the options and pre-processes them to ensure we aren't trying to access non-existent